Enable test for missing ~A section data.
The purpose of this test is to validate that wellio.js doesn't crash or
panic when the ~A section has no data.

The test code was fixed to check the status of the internal data
representatin for the CURVES (~A) section.  No changes were needed to
the main wellio.js code.