Issue 12: Handle wrapped data properly
This change removes 'skip' from the wrapped data test.
It also checks if the data is wrapped. If the data is wrapped,
then data elements will be accumulated for till we have the same
number as columns/curve names (from the 'CURVE INFORMATION' block).
This should insure that we have the right number of data elements per
'row'.