Wellio-Schema: initial commit
This branch and its commits are for exploring and discussing the
usefulness of a possible Wellio json-schema. This commit contains an
initial wellio.schema.json file that specs out a minimal LAS Version
section. The added wellio_sample_2.0.json is the wellio json based on
sample_2.0.las.  The lasio json files are renamed to have 'lasio_' as a
prefix to distinguish them from wellio json files.  This branch also
adds the ajv-cli tool to the npm package.json devDependencies.  The
ajv-cli tool is a JavaScript-based json-schema validation tool.